feat: Initial project commit
Here I added my main python file for the bikeshare project and created
a .gitignore file to store all three provided city data documents that
were used to interact with the project output.

I have not made any changes to resolve my coding, so I don't have a
problem that needs to be resolved and explained as of now.