refactor: var changes and line removals
Used find and replace to update two variables from df to dataframe and
t to time_parsed, so it will be easier to understand what the variables
stand for and removed a few empty lines that were unnecessary to make
more lean and clean.

No bugs needed to be fixed, so I have no number of issues to list.